requirements.txt
streamlit==1.27.0
pandas==2.1.0
google-cloud-bigquery==3.11.4 